backend/requirements.txt
pandas
sqlalchemy
pymysql
pytesseract
pillow
FastAPI
Uvicorn
mysql-connector-python
pdfplumber
Pillow
python-multipart
